@@ -10,40 +14,129 @@
   "keywords": [
     "uuid",
     "guid",
-    "rfc4122"
+    "rfc4122",
+    "rfc9562"
   ],
   "license": "MIT",
   "bin": {
-    "uuid": "./bin/uuid"
+    "uuid": "dist/bin/uuid"
   },
+  "sideEffects": false,
+  "main": "./dist/index.js",
+  "exports": {
+    ".": {
+      "node": {
+        "module": "./dist/esm-node/index.js",
+        "require": "./dist/index.js",
+        "import": "./wrapper.mjs"
+      },
+      "browser": {
+        "import": "./dist/esm-browser/index.js",
+        "require": "./dist/commonjs-browser/index.js"
+      },
+      "default": "./dist/esm-browser/index.js"
+    },
+    "./package.json": "./package.json"
+  },
+  "module": "./dist/esm-node/index.js",
+  "browser": {
+    "./dist/esm-node/index.js": "./dist/esm-browser/index.js",
+    "./dist/md5.js": "./dist/md5-browser.js",
+    "./dist/native.js": "./dist/native-browser.js",
+    "./dist/rng.js": "./dist/rng-browser.js",
+    "./dist/sha1.js": "./dist/sha1-browser.js"
+  },
+  "files": [
+    "CHANGELOG.md",
+    "CONTRIBUTING.md",
+    "LICENSE.md",
+    "README.md",
+    "dist",
+    "wrapper.mjs"
+  ],
   "devDependencies": {
-    "@commitlint/cli": "~8.2.0",
-    "@commitlint/config-conventional": "~8.2.0",
-    "eslint": "~6.4.0",
-    "husky": "~3.0.5",
-    "mocha": "6.2.0",
-    "runmd": "1.2.1",
-    "standard-version": "7.0.0"
+    "@babel/cli": "7.24.6",
+    "@babel/core": "7.24.6",
+    "@babel/eslint-parser": "7.24.6",
+    "@babel/plugin-syntax-import-attributes": "7.24.6",
+    "@babel/preset-env": "7.24.6",
+    "@commitlint/cli": "19.3.0",
+    "@commitlint/config-conventional": "19.2.2",
+    "@wdio/browserstack-service": "7.16.10",
+    "@wdio/cli": "7.16.10",
+    "@wdio/jasmine-framework": "7.16.6",
+    "@wdio/local-runner": "7.16.10",
+    "@wdio/spec-reporter": "7.16.9",
+    "@wdio/static-server-service": "7.16.6",
+    "bundlewatch": "0.3.3",
+    "eslint": "9.4.0",
+    "eslint-plugin-prettier": "5.1.3",
+    "globals": "15.3.0",
+    "husky": "9.0.11",
+    "jest": "29.7.0",
+    "lint-staged": "15.2.5",
+    "neostandard": "0.5.1",
+    "npm-run-all": "4.1.5",
+    "optional-dev-dependency": "2.0.1",
+    "prettier": "3.3.0",
+    "random-seed": "0.3.0",
+    "runmd": "1.3.9",
+    "standard-version": "9.5.0"
+  },
+  "optionalDevDependencies": {
+    "@wdio/browserstack-service": "7.16.10",
+    "@wdio/cli": "7.16.10",
+    "@wdio/jasmine-framework": "7.16.6",
+    "@wdio/local-runner": "7.16.10",
+    "@wdio/spec-reporter": "7.16.9",
+    "@wdio/static-server-service": "7.16.6"
   },
   "scripts": {
-    "lint": "eslint .",
-    "test": "npm run lint && mocha test/test.js",
+    "examples:browser:webpack:build": "cd examples/browser-webpack && npm install && npm run build",
+    "examples:browser:rollup:build": "cd examples/browser-rollup && npm install && npm run build",
+    "examples:node:commonjs:test": "cd examples/node-commonjs && npm install && npm test",
+    "examples:node:esmodules:test": "cd examples/node-esmodules && npm install && npm test",
+    "examples:node:jest:test": "cd examples/node-jest && npm install && npm test",
+    "prepare": "husky install",
+    "lint": "npm run eslint:check && npm run prettier:check",
+    "eslint:check": "eslint src/ test/ examples/ *.js",
+    "eslint:fix": "eslint --fix src/ test/ examples/ *.js",
+    "pretest": "npm run build",
+    "test": "BABEL_ENV=commonjsNode node --throw-deprecation node_modules/.bin/jest test/unit/",
+    "test:matching": "BABEL_ENV=commonjsNode node --throw-deprecation node_modules/.bin/jest test/unit/ -t",
+    "pretest:browser": "optional-dev-dependency && npm run build && npm-run-all --parallel examples:browser:**",
+    "test:browser": "wdio run ./wdio.conf.js",
+    "pretest:node": "npm run build",
+    "test:node": "npm-run-all --parallel examples:node:**",
+    "test:pack": "./scripts/testpack.sh",
+    "pretest:benchmark": "npm run build",
+    "test:benchmark": "cd examples/benchmark && HUSKY=0 npm install && npm test",
+    "prettier:check": "prettier --check .",
+    "prettier:fix": "prettier --write .",
+    "bundlewatch": "npm run pretest:browser && bundlewatch --config bundlewatch.config.json",
     "md": "runmd --watch --output=README.md README_js.md",
-    "release": "standard-version",
-    "prepare": "runmd --output=README.md README_js.md"
-  },
-  "browser": {
-    "./lib/rng.js": "./lib/rng-browser.js",
-    "./lib/sha1.js": "./lib/sha1-browser.js",
-    "./lib/md5.js": "./lib/md5-browser.js"
+    "docs": "npm run build && npx runmd --output=README.md README_js.md",
+    "docs:diff": "npm run docs && git diff --quiet README.md",
+    "build": "./scripts/build.sh",
+    "prepack": "npm run build",
+    "release": "standard-version --no-verify"
   },
   "repository": {
     "type": "git",
-    "url": "https://github.com/uuidjs/uuid.git"
+    "url": "git+https://github.com/uuidjs/uuid.git"
   },
-  "husky": {
-    "hooks": {
-      "commit-msg": "commitlint -E HUSKY_GIT_PARAMS"
+  "lint-staged": {
+    "*": [
+      "prettier --no-error-on-unmatched-pattern --write"
+    ],
+    "*.{js,jsx}": [
+      "eslint --no-error-on-unmatched-pattern --fix"
+    ]
+  },
+  "standard-version": {
+    "scripts": {
+      "postchangelog": "prettier --write CHANGELOG.md"
     }
-  }
+  },
+  "packageManager": "npm@10.8.1+sha256.b8807aebb9656758e2872fa6e7c564b506aa2faa9297439a478d471d2fe32483"
 }
